requirements.txt
gunicorn
flask
psycopg2-binary
python-dotenv